test(reporter-dom): move to uvu
