+++ packages/reporter-react/package.json
@@ -37,8 +37,8 @@
     "build": "pnpm prebuild && rollup -c",
     "dev": "rollup -cw",
     "prepublishOnly": "pnpm run build",
-    "test": "jest",
-    "test:ci": "jest --ci --coverage"
+    "test": "uvu -r tsm -r global-jsdom/register tests -i common",
+    "test:ci": "nyc -n src pnpm test"
   },
   "dependencies": {
     "@felte/common": "workspace:*"
